+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Changes from v0.2 to v0.3
 * Fix warnings in boost::random stuff on some systems (Issue 250)
 
 * Minor changes in the python interface to the outputs of the moments and shape estimation routines
-  (the HSMShapeData class).  (Issue 296)
+  (the HSMShapeData class).  (Issue 296, 316, 332)
 
 * On systems where a different C++ compiler was used for GalSim and for python, C++ exceptions show
   up with a non-informative error message.  While this is not easily fixable, there is now a test
@@ -40,3 +40,8 @@ Changes from v0.2 to v0.3
   leaks in the C++ side of GalSim. (Issue 327)
 
 * Enable copying Images of different types. (Issue 327)
+
+* The moments and PSF correction code was updated to use the Image class and TMV. The python
+  interface to this software was also updated so that it can handle weight and bad pixel maps for
+  the input Images.  Finally, an optimization was introduced that typically speeds up these routines
+  by a factor of 2-3.  (Issues 331, 332)